try-with-resources: session auto close

+++ erp-logics/src/main/java/lsfusion/erp/region/by/certificate/declaration/ImportTNVEDClassifierActionProperty.java
@@ -35,8 +35,8 @@ public class ImportTNVEDClassifierActionProperty extends ScriptingActionProperty
 
         try {
 
-            Object countryBelarus = findProperty("countrySID").read(context.getSession(), new DataObject("112", StringClass.get(3)));
-            findProperty("defaultCountry").change(countryBelarus, context.getSession());
+            Object countryBelarus = findProperty("countrySID").read(context, new DataObject("112", StringClass.get(3)));
+            findProperty("defaultCountry").change(countryBelarus, context);
             context.getSession().apply(context);
 
             CustomStaticFormatFileClass valueClass = CustomStaticFormatFileClass.get(false, false, "Файлы DBF", "DBF");
@@ -49,44 +49,49 @@ public class ImportTNVEDClassifierActionProperty extends ScriptingActionProperty
                     importParents(context, file);
                 }
             }
-        } catch (xBaseJException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        } catch (ScriptingErrorLog.SemanticErrorException e) {
+        } catch (xBaseJException | IOException | ScriptingErrorLog.SemanticErrorException e) {
             throw new RuntimeException(e);
         }
     }
 
     private void importGroups(ExecutionContext<ClassPropertyInterface> context, byte[] fileBytes) throws IOException, xBaseJException, ScriptingErrorLog.SemanticErrorException, SQLException, SQLHandledException {
 
-        File tempFile = File.createTempFile("tempTnved", ".dbf");
-        IOUtils.putFileBytes(tempFile, fileBytes);
+        List<List<Object>> data = new ArrayList<>();
 
-        DBF file = new DBF(tempFile.getPath());
+        File tempFile = null;
+        DBF dbfFile = null;
+        try {
 
-        List<List<Object>> data = new ArrayList<List<Object>>();
+            tempFile = File.createTempFile("tempTnved", ".dbf");
+            IOUtils.putFileBytes(tempFile, fileBytes);
 
-        BigDecimal defaultVAT = new BigDecimal(20);
-        Date defaultDate = new Date(2001 - 1900, 0, 1);
+            dbfFile = new DBF(tempFile.getPath());
+            int recordCount = dbfFile.getRecordCount();
 
-        int recordCount = file.getRecordCount();
-        for (int i = 1; i <= recordCount; i++) {
-            file.read();
+            BigDecimal defaultVAT = new BigDecimal(20);
+            Date defaultDate = new Date(2001 - 1900, 0, 1);
 
-            String groupID = new String(file.getField("KOD").getBytes(), "Cp866").trim();
-            String name = new String(file.getField("NAIM").getBytes(), "Cp866").trim();
-            String extraName = new String(file.getField("KR_NAIM").getBytes(), "Cp866").trim();
 
-            Boolean hasCode = true;
-            if (groupID.equals("··········")) {
-                groupID = "-" + i;
-                hasCode = null;
+            for (int i = 1; i <= recordCount; i++) {
+                dbfFile.read();
+
+                String groupID = new String(dbfFile.getField("KOD").getBytes(), "Cp866").trim();
+                String name = new String(dbfFile.getField("NAIM").getBytes(), "Cp866").trim();
+                String extraName = new String(dbfFile.getField("KR_NAIM").getBytes(), "Cp866").trim();
+
+                Boolean hasCode = true;
+                if (groupID.equals("··········")) {
+                    groupID = "-" + i;
+                    hasCode = null;
+                }
+                data.add(Arrays.asList((Object) groupID, name + extraName, i, "БЕЛАРУСЬ", hasCode, defaultVAT, defaultDate));
             }
-            data.add(Arrays.asList((Object) groupID, name + extraName, i, "БЕЛАРУСЬ", hasCode, defaultVAT, defaultDate));
+        } finally {
+            if(dbfFile != null)
+                dbfFile.close();
+            if(tempFile != null)
+            tempFile.delete();
         }
-        file.close();
-        tempFile.delete();
         
         ImportField codeCustomsGroupField = new ImportField(findProperty("codeCustomsGroup"));
         ImportField nameCustomsGroupField = new ImportField(findProperty("nameCustomsGroup"));
@@ -98,9 +103,9 @@ public class ImportTNVEDClassifierActionProperty extends ScriptingActionProperty
 
         ImportKey<?> customsGroupKey = new ImportKey((CustomClass) findClass("CustomsGroup"), findProperty("customsGroupCode").getMapping(codeCustomsGroupField));
         ImportKey<?> customsZoneKey = new ImportKey((CustomClass) findClass("CustomsZone"), findProperty("customsZoneName").getMapping(nameCustomsZoneField));
-        ImportKey<?> VATKey = new ImportKey((ConcreteCustomClass) findClass("Range"), findProperty("valueCurrentVATDefaultValue").getMapping(vatField));
+        ImportKey<?> VATKey = new ImportKey((CustomClass) findClass("Range"), findProperty("valueCurrentVATDefaultValue").getMapping(vatField));
 
-        List<ImportProperty<?>> properties = new ArrayList<ImportProperty<?>>();
+        List<ImportProperty<?>> properties = new ArrayList<>();
         properties.add(new ImportProperty(codeCustomsGroupField, findProperty("codeCustomsGroup").getMapping(customsGroupKey)));
         properties.add(new ImportProperty(nameCustomsGroupField, findProperty("nameCustomsGroup").getMapping(customsGroupKey)));
         properties.add(new ImportProperty(numberCustomsGroupField, findProperty("numberCustomsGroup").getMapping(customsGroupKey)));
@@ -112,12 +117,12 @@ public class ImportTNVEDClassifierActionProperty extends ScriptingActionProperty
         ImportTable table = new ImportTable(Arrays.asList(codeCustomsGroupField, nameCustomsGroupField,
                 numberCustomsGroupField, nameCustomsZoneField, hasCodeCustomsGroupField, vatField, dateField), data);
 
-        DataSession session = context.createSession();
-        IntegrationService service = new IntegrationService(session, table,
-                Arrays.asList(customsGroupKey, customsZoneKey, VATKey), properties);
-        service.synchronize(true, false);
-        session.apply(context);
-        session.close();
+        try (DataSession session = context.createSession()) {
+            IntegrationService service = new IntegrationService(session, table,
+                    Arrays.asList(customsGroupKey, customsZoneKey, VATKey), properties);
+            service.synchronize(true, false);
+            session.apply(context);
+        }
     }
 
     private void importParents(ExecutionContext<ClassPropertyInterface> context, byte[] fileBytes) throws IOException, xBaseJException, ScriptingErrorLog.SemanticErrorException, SQLException, SQLHandledException {
@@ -127,8 +132,8 @@ public class ImportTNVEDClassifierActionProperty extends ScriptingActionProperty
 
         DBF file = new DBF(tempFile.getPath());
 
-        List<List<Object>> data = new ArrayList<List<Object>>();
-        List<String> groupIDsList = new ArrayList<String>();
+        List<List<Object>> data = new ArrayList<>();
+        List<String> groupIDsList = new ArrayList<>();
         int recordCount = file.getRecordCount();
         for (int i = 1; i <= recordCount; i++) {
             file.read();
@@ -159,17 +164,17 @@ public class ImportTNVEDClassifierActionProperty extends ScriptingActionProperty
         ImportKey<?> customsGroupKey = new ImportKey((CustomClass) findClass("CustomsGroup"), findProperty("customsGroupCode").getMapping(groupIDField));
         ImportKey<?> parentCustomsGroupKey = new ImportKey((CustomClass) findClass("CustomsGroup"), findProperty("customsGroupCode").getMapping(parentIDField));
 
-        List<ImportProperty<?>> properties = new ArrayList<ImportProperty<?>>();
+        List<ImportProperty<?>> properties = new ArrayList<>();
         properties.add(new ImportProperty(parentIDField, findProperty("parentCustomsGroup").getMapping(customsGroupKey),
                 object(findClass("CustomsGroup")).getMapping(parentCustomsGroupKey)));
 
         ImportTable table = new ImportTable(Arrays.asList(groupIDField, parentIDField), data);
 
-        DataSession session = context.createSession();
-        IntegrationService service = new IntegrationService(session, table,
-                Arrays.asList(customsGroupKey, parentCustomsGroupKey), properties);
-        service.synchronize(true, false);
-        session.apply(context);
-        session.close();
+        try (DataSession session = context.createSession()) {
+            IntegrationService service = new IntegrationService(session, table,
+                    Arrays.asList(customsGroupKey, parentCustomsGroupKey), properties);
+            service.synchronize(true, false);
+            session.apply(context);
+        }
     }
 }
