+++ erp-logics/src/main/java/lsfusion/erp/region/by/certificate/declaration/ImportTNVEDCustomsExceptionsActionProperty.java
@@ -4,7 +4,6 @@ import lsfusion.base.IOUtils;
 import lsfusion.base.col.interfaces.immutable.ImMap;
 import lsfusion.base.col.interfaces.immutable.ImOrderMap;
 import lsfusion.base.col.interfaces.immutable.ImRevMap;
-import lsfusion.server.classes.ConcreteCustomClass;
 import lsfusion.server.classes.CustomClass;
 import lsfusion.server.classes.CustomStaticFormatFileClass;
 import lsfusion.server.classes.DateClass;
@@ -52,13 +51,7 @@ public class ImportTNVEDCustomsExceptionsActionProperty extends ScriptingActionP
                 }
             }
 
-        } catch (xBaseJException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        } catch (ScriptingErrorLog.SemanticErrorException e) {
-            throw new RuntimeException(e);
-        } catch (ParseException e) {
+        } catch (xBaseJException | IOException | ScriptingErrorLog.SemanticErrorException | ParseException e) {
             throw new RuntimeException(e);
         }
     }
@@ -67,9 +60,9 @@ public class ImportTNVEDCustomsExceptionsActionProperty extends ScriptingActionP
 
         List<List<Object>> data = importVATExceptionFromDBF(context, fileBytes);
 
-        List<ImportProperty<?>> props = new ArrayList<ImportProperty<?>>();
-        List<ImportField> fields = new ArrayList<ImportField>();
-        List<ImportKey<?>> keys = new ArrayList<ImportKey<?>>();
+        List<ImportProperty<?>> props = new ArrayList<>();
+        List<ImportField> fields = new ArrayList<>();
+        List<ImportKey<?>> keys = new ArrayList<>();
 
         ImportField codeCustomsGroupField = new ImportField(findProperty("codeCustomsGroup"));
         ImportKey<?> customsGroupKey = new ImportKey((CustomClass) findClass("CustomsGroup"),
@@ -89,7 +82,7 @@ public class ImportTNVEDCustomsExceptionsActionProperty extends ScriptingActionP
         fields.add(nameVATCustomsExceptionField);
 
         ImportField rangeField = new ImportField(findProperty("dataValueSupplierVATCustomsGroupDate"));
-        ImportKey<?> rangeKey = new ImportKey((ConcreteCustomClass) findClass("Range"),
+        ImportKey<?> rangeKey = new ImportKey((CustomClass) findClass("Range"),
                 findProperty("valueCurrentVATDefaultValue").getMapping(rangeField));
         props.add(new ImportProperty(rangeField, findProperty("rangeVATCustomsException").getMapping(VATCustomsExceptionKey),
                 object(findClass("Range")).getMapping(rangeKey)));
@@ -107,46 +100,52 @@ public class ImportTNVEDCustomsExceptionsActionProperty extends ScriptingActionP
 
         ImportTable table = new ImportTable(fields, data);
 
-        DataSession session = context.createSession();
-        IntegrationService service = new IntegrationService(session, table, keys, props);
-        service.synchronize(true, false);
-        session.apply(context);
-        session.close();
+        try (DataSession session = context.createSession()) {
+            IntegrationService service = new IntegrationService(session, table, keys, props);
+            service.synchronize(true, false);
+            session.apply(context);
+        }
     }
 
     private List<List<Object>> importVATExceptionFromDBF(ExecutionContext context, byte[] fileBytes) throws IOException, xBaseJException, ParseException, ScriptingErrorLog.SemanticErrorException, SQLException, SQLHandledException {
 
+        List<List<Object>> data = new ArrayList<>();
+        Map<String, List<Object>> dataVATMap = new HashMap<>();
         Set<String> tnvedSet = getTNVEDSet(context);
 
-        File tempFile = File.createTempFile("tempTnved", ".dbf");
-        IOUtils.putFileBytes(tempFile, fileBytes);
+        File tempFile = null;
+        DBF dbfFile = null;
+        try {
+            tempFile = File.createTempFile("tempTnved", ".dbf");
+            IOUtils.putFileBytes(tempFile, fileBytes);
 
-        DBF file = new DBF(tempFile.getPath());
+            dbfFile = new DBF(tempFile.getPath());
+            int recordCount = dbfFile.getRecordCount();
 
-        List<List<Object>> data = new ArrayList<List<Object>>();
-        Map<String, List<Object>> dataVATMap = new HashMap<String, List<Object>>();
+            for (int i = 1; i <= recordCount; i++) {
+                dbfFile.read();
 
-        int recordCount = file.getRecordCount();
-        for (int i = 1; i <= recordCount; i++) {
-            file.read();
+                Integer type = Integer.parseInt(new String(dbfFile.getField("PP").getBytes(), "Cp866").trim());
+                String codeCustomsGroup = new String(dbfFile.getField("G33").getBytes(), "Cp866").trim();
+                String name = new String(dbfFile.getField("TEXT").getBytes(), "Cp866").trim();
+                BigDecimal stav1 = new BigDecimal(new String(dbfFile.getField("STAV1").getBytes(), "Cp866").trim());
+                Date dateFrom = new Date(DateUtils.parseDate(new String(dbfFile.getField("DATE1").getBytes(), "Cp866").trim(), new String[]{"yyyyMMdd"}).getTime());
+                Date dateTo = new Date(DateUtils.parseDate(new String(dbfFile.getField("DATE2").getBytes(), "Cp866").trim(), new String[]{"yyyyMMdd"}).getTime());
 
-            Integer type = Integer.parseInt(new String(file.getField("PP").getBytes(), "Cp866").trim());
-            String codeCustomsGroup = new String(file.getField("G33").getBytes(), "Cp866").trim();
-            String name = new String(file.getField("TEXT").getBytes(), "Cp866").trim();
-            BigDecimal stav1 = new BigDecimal(new String(file.getField("STAV1").getBytes(), "Cp866").trim());
-            Date dateFrom = new Date(DateUtils.parseDate(new String(file.getField("DATE1").getBytes(), "Cp866").trim(), new String[]{"yyyyMMdd"}).getTime());
-            Date dateTo = new Date(DateUtils.parseDate(new String(file.getField("DATE2").getBytes(), "Cp866").trim(), new String[]{"yyyyMMdd"}).getTime());
-
-            if (type.equals(4)) {
-                if(codeCustomsGroup.length()==10)
-                    data.add(Arrays.asList((Object) codeCustomsGroup, codeCustomsGroup + String.valueOf(dateTo) + name, name, stav1, dateFrom, dateTo));
-                else
-                    dataVATMap.put(codeCustomsGroup, Arrays.asList((Object) codeCustomsGroup, codeCustomsGroup + String.valueOf(dateTo) + name, name, stav1, dateFrom, dateTo));
+                if (type.equals(4)) {
+                    if (codeCustomsGroup.length() == 10)
+                        data.add(Arrays.asList((Object) codeCustomsGroup, codeCustomsGroup + String.valueOf(dateTo) + name, name, stav1, dateFrom, dateTo));
+                    else
+                        dataVATMap.put(codeCustomsGroup, Arrays.asList((Object) codeCustomsGroup, codeCustomsGroup + String.valueOf(dateTo) + name, name, stav1, dateFrom, dateTo));
+                }
             }
+        } finally {
+            if(dbfFile != null)
+                dbfFile.close();
+            if(tempFile != null)
+                tempFile.delete();
         }
-        file.close();
-        tempFile.delete();
-
+        
         for (Map.Entry<String, List<Object>> entry : dataVATMap.entrySet()) {
             for(String tnved : tnvedSet) {
                 if( tnved.startsWith(entry.getKey())) {
@@ -161,12 +160,12 @@ public class ImportTNVEDCustomsExceptionsActionProperty extends ScriptingActionP
 
     private Set<String> getTNVEDSet(ExecutionContext context) throws ScriptingErrorLog.SemanticErrorException, SQLException, SQLHandledException {
 
-        Set<String> tnvedSet = new HashSet<String>();
+        Set<String> tnvedSet = new HashSet<>();
 
         LCP<?> isCustomsGroup = is(findClass("CustomsGroup"));
         ImRevMap<Object, KeyExpr> keys = (ImRevMap<Object, KeyExpr>) isCustomsGroup.getMapKeys();
         KeyExpr key = keys.singleValue();
-        QueryBuilder<Object, Object> query = new QueryBuilder<Object, Object>(keys);
+        QueryBuilder<Object, Object> query = new QueryBuilder<>(keys);
         query.addProperty("codeCustomsGroup", findProperty("codeCustomsGroup").getExpr(context.getModifier(), key));
         query.and(isCustomsGroup.getExpr(key).getWhere());
         ImOrderMap<ImMap<Object, Object>, ImMap<Object, Object>> result = query.execute(context);
